+++ scripts/pdf-editable.mjs
@@ -31,8 +31,9 @@ SOFTWARE.
 
 import { PDFCONFIG } from './pdf-config.mjs'
 
-let pdf2actormap;                   // key = pdf field name, value = actor field name
-let actor2pdfviewer = new Map();    // key = actorid, value = container
+let map_pdf2actor;                   // key = pdf field name, value = actor field name
+let map_pdf2item;                    // key = pdf field name, value = item  field name
+let document2pdfviewer = new Map();    // key = document uuid, value = container
 
 // Function to convert Object into a string whilst keeping the functions
 function Obj2String(obj) {
@@ -53,38 +54,36 @@ function Obj2String(obj) {
     return ret;
 }
 
+// fromUuidSync(id) - needs to be prefixed with "Actor." or "Item."
+
 /**
- * Copy all the data from the specified document (Actor) to the fields on the PDF sheet (container)
+ * Copy all the data from the specified Document (Actor/Item) to the fields on the PDF sheet (container)
  * @param {PDFPageView} pdfpageview  
- * @param {Actor} actor 
+ * @param {Document} document An Actor or Item
  */
-async function setFormFromActor(pdfpageview, actor) {
-    console.debug(`${PDFCONFIG.MODULE_NAME}: setting values for '${actor.name}'`)
-    let container = pdfpageview.div;
+async function setFormFromDocument(pdfpageview, document) {
+    //console.debug(`${PDFCONFIG.MODULE_NAME}: setting values for '${document.name}'`)
+    let flags = document.getFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_FIELDTEXT) || {};
 
-    let flags = game.settings.get(PDFCONFIG.MODULE_NAME, PDFCONFIG.STORE_FIELDS_ON_ACTOR) 
-        ? ( actor.getFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_FIELDTEXT) || {} )
-        : undefined;
-
-    const inputs = container.querySelectorAll('input,textarea');
-
-    // Set values from the module FLAG on the actor.
+    // Set values from the module FLAG on the Document.
+    const inputs = pdfpageview.div.querySelectorAll('input,textarea');
+    const mapping = (document instanceof Actor) ? map_pdf2actor : map_pdf2item;
     for (let elem of inputs) {
         // don't modify disabled (readonly) fields
         if (elem.disabled) continue; // DISABLED
 
         // Get required value, either from the FLAGS or from a field on the Actor
+        const docfield = mapping && (mapping[elem.name] || mapping[elem.id]);
         let value;
-        if (flags) {
+        if (!docfield) {
             value = flags[elem.name];
             if (value === undefined) value = flags[elem.id];
         } else {
-            const actorfield = pdf2actormap[(elem.name||elem.id).trim()];
-            if (actorfield instanceof Object && actorfield.getValue) {
-                value = actorfield.getValue(actor);
-                if (!actorfield.setValue) elem.readOnly = true;
-            } else if (typeof actorfield === 'string') {
-                value = foundry.utils.getProperty(actor, actorfield);
+            if (docfield instanceof Object && docfield.getValue) {
+                value = docfield.getValue(document);
+                if (!docfield.setValue) elem.readOnly = true;
+            } else if (typeof docfield === 'string') {
+                value = foundry.utils.getProperty(document, docfield);
                 if (typeof value === 'number') value = '' + value;
             }
         }
@@ -103,45 +102,58 @@ async function setFormFromActor(pdfpageview, actor) {
 
 /**
  * Handle the event which fired when the user changed a value in a field.
- * @param {*} actor 
- * @param {*} inputid 
- * @param {*} inputname 
- * @param {*} value 
+ * @param {Document} document An Actor or Item
+ * @param {string} inputid The PDF ID of the field
+ * @param {string} inputname  The PDF NAME of the field
+ * @param {string} value The value to be stored in the document
  */
-function modifyActor(actor, inputid, inputname, value) {
-    if (game.settings.get(PDFCONFIG.MODULE_NAME, PDFCONFIG.STORE_FIELDS_ON_ACTOR)) {
-        // Copy the modified field to the MODULE FLAG on the Actor
-        let flags = actor.getFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_FIELDTEXT);
+function modifyDocument(document, inputid, inputname, value) {
+    // Copy the modified field to the corresponding field in the Document
+    const mapping = (document instanceof Actor) ? map_pdf2actor : map_pdf2item;
+    let docfield = mapping && (mapping[inputname] || mapping[inputid]);
+    if (!docfield) {
+        console.debug(`${PDFCONFIG.MODULE_NAME}: unmapped PDF field: NAME '${inputname}' ID '${inputid}', = '${value}'`);
+        // Copy the modified field to the MODULE FLAG in the Document
+        let flags = document.getFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_FIELDTEXT);
         if (!flags || !(flags instanceof Object)) flags = {};
         let docfield = inputname || inputid;
         if (flags[docfield] === value) return;
         flags[docfield] = value;
-        console.debug(`FLAGS for '${actor.name}': '${inputname || inputid}' = '${value}'`);
-        actor.setFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_FIELDTEXT, flags);
-    } else {
-        // Copy the modified field to the corresponding field on the Actor
-        let docfield = pdf2actormap[inputname.trim()] || pdf2actormap[inputid.trim()];
-        if (!docfield)
-            console.debug(`${PDFCONFIG.MODULE_NAME}: unmapped PDF field:\nID '${inputid}', NAME '${inputname}' = '${value}'`)
-        else if (typeof docfield === 'string') {
-            console.debug(`STRING for '${actor.name}': '${docfield}' = '${value}'`);
-            actor.update({ [docfield]: value });
-        } else if (docfield.setValue) {
-            console.debug(`SETVALUE for '${actor.name}': 'calculated field = '${value}'`);
-            docfield.setValue(actor, value);
-        }
+        console.debug(`Storing hidden value on '${document.name}': '${docfield}' = '${value}'`);
+        document.setFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_FIELDTEXT, flags);        
+    } else if (typeof docfield === 'string') {
+        console.debug(`Updating '${document.name}'['${docfield}'] = '${value}'`);
+        if (getProperty(document, docfield) === value) return;
+        document.update({ [docfield]: value });
+    } else if (docfield.setValue) {
+        console.debug(`Calling setValue on field for '${document.name}' with '${value}'`);
+        docfield.setValue(document, value);
     }
 }
 
 /**
  * Handle updates to actors which are present in one of the open PDF sheets.
  */
-Hooks.on('updateActor', async function(actordoc, change, options, userId) {
-    let pdfviewer = actor2pdfviewer.get(actordoc.id);
+async function update_document(document, change, options, userId) {
+    let pdfviewer = document2pdfviewer.get(document.uuid);
     if (!pdfviewer) return;
-    console.log(`${PDFCONFIG.MODULE_NAME}: Update of actor ${actordoc.name}`);
-    setFormFromActor(pdfviewer, actordoc);
-})
+    setFormFromDocument(pdfviewer, document);
+}
+Hooks.on('updateActor', update_document)
+Hooks.on('updateItem',  update_document)
+
+/**
+ * Handle a displayed item being deleted, break the connection to the displayed PDF page.
+ */
+async function delete_document(document, options, userId) {
+    console.log(`Item ${document.uuid} was deleted`)
+    let pdfviewer = document2pdfviewer.get(document.uuid);
+    if (!pdfviewer) return;
+    document2pdfviewer.delete(document.uuid);
+}
+Hooks.on('deleteActor', delete_document)
+Hooks.on('deleteItem',  delete_document)
+
 
 /**
  * Called from renderJournalPDFPageSheet
@@ -149,43 +161,46 @@ Hooks.on('updateActor', async function(actordoc, change, options, userId) {
  */
 export async function initEditor(sheet, html, data) {
 
-    const actorid = sheet.object.getFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_CODE);
-    if (!actorid) return;
-    const actor = game.actors.get(actorid);
-    if (!actor) return;
+    const pdfcode = sheet.object.getFlag(PDFCONFIG.MODULE_NAME, PDFCONFIG.FLAG_CODE);
+    if (!pdfcode) return;
+    const document = fromUuidSync(pdfcode) || game.actors.get(pdfcode) || game.actors.get(pdfcode);
+    if (!document) return;
 
-    // Load system JSON if not already done (but only if needing to update real Actor fields)
-    if (!game.settings.get(PDFCONFIG.MODULE_NAME, PDFCONFIG.STORE_FIELDS_ON_ACTOR)) {
-        // Always reload the MAP on opening the window (in case it has changed since last time)
-        let flag = game.settings.get(PDFCONFIG.MODULE_NAME, PDFCONFIG.ACTOR_CONFIG);
-        if (flag) pdf2actormap = eval(`(${flag})`);
+    // Always reload the MAP on opening the window (in case it has changed since last time)
+    let mapping = game.settings.get(PDFCONFIG.MODULE_NAME, PDFCONFIG.ACTOR_CONFIG);
+    if (mapping) map_pdf2actor = eval(`(${mapping})`);
+    mapping = game.settings.get(PDFCONFIG.MODULE_NAME, PDFCONFIG.ITEM_CONFIG);
+    if (mapping) map_pdf2item = eval(`(${mapping})`);
 
-        if (!pdf2actormap) {
-            const module = await import(`../systems/${game.system.id}.mjs`)
-                .then(module => {
-                    pdf2actormap = module.namemap;
-                    let mapping = Obj2String(pdf2actormap);
-                    game.settings.set(PDFCONFIG.MODULE_NAME, PDFCONFIG.ACTOR_CONFIG, mapping);
-                })
-                .catch(error => console.warn(`${PDFCONFIG.MODULE_NAME}: Failed to find the PDF mapping file for 'systems/${game.system.id}.mjs'. Form-Filled PDFs will not work`))
-        }
-        if (!pdf2actormap) return;
+    const systemfile = `/systems/${game.system.id}.mjs`;
+    if (!map_pdf2actor && await srcExists(`modules/${PDFCONFIG.MODULE_NAME}${systemfile}`)) {
+        const module = await import(`..${systemfile}`)
+            .then(module => {
+                map_pdf2actor = module.actormap;
+                map_pdf2item  = module.itemmap;
+                game.settings.set(PDFCONFIG.MODULE_NAME, PDFCONFIG.ACTOR_CONFIG, Obj2String(map_pdf2actor));
+                game.settings.set(PDFCONFIG.MODULE_NAME, PDFCONFIG.ITEM_CONFIG,  Obj2String(map_pdf2item));
+            })
+            .catch(error => console.warn(`${PDFCONFIG.MODULE_NAME}: Failed to find the PDF mapping file for 'systems/${game.system.id}.mjs'.`))
+    } else {
+        if (!map_pdf2actor) map_pdf2actor = {};
+        if (!map_pdf2item)  map_pdf2item  = {};
     }
 
     // Wait for the IFRAME to appear in the window before any further initialisation
     $('iframe').on('load', async (event) => {
-        console.debug(`${PDFCONFIG.MODULE_NAME}: PDF frame loaded for '${sheet.object.name}'`);
+        //console.debug(`${PDFCONFIG.MODULE_NAME}: PDF frame loaded for '${sheet.object.name}'`);
 
         // Wait for the PDFViewer to be fully initialized
         const contentWindow = event.target.contentWindow;
-        // Keep ACTOR->container mapping only while the window is open
+        // Keep DOCUMENT->container mapping only while the window is open
         contentWindow.addEventListener('unload', (event) => {
-            actor2pdfviewer.delete(actorid);
+            document2pdfviewer.delete(document.uuid);
         })
 
         const pdfviewerapp = contentWindow.PDFViewerApplication;
         await pdfviewerapp.initializedPromise;
-        console.debug(`${PDFCONFIG.MODULE_NAME}: PDFJS initialized`);  // but document still NOT loaded
+        //console.debug(`${PDFCONFIG.MODULE_NAME}: PDFJS initialized`);  // but document still NOT loaded
 
         // "change" events are NOT sent for fields which have JS Actions attached to them,
         // so we have to attach to the PDF viewer's dispatcheventinsandbox events.
@@ -193,17 +208,15 @@ export async function initEditor(sheet, html, data) {
         // 'willCommit' is for text fields
         pdfviewerapp.eventBus.on('dispatcheventinsandbox', (event) => {
             if (event.detail.name == 'Action' || event.detail.willCommit) {
-                console.log(`dispatcheventinsandbox: id = '${event.detail.id}', field = '${event.source.data.fieldName}', value = '${event.detail.value}'`);
-                modifyActor(actor, event.detail.id, event.source?.data.fieldName, event.detail.value);
+                //console.log(`dispatcheventinsandbox: id = '${event.detail.id}', field = '${event.source.data.fieldName}', value = '${event.detail.value}'`);
+                modifyDocument(document, event.detail.id, event.source?.data.fieldName, event.detail.value);
             }
         })
 
-        // Wait for the AnnotationLayer to get drawn before populating all the fields with data from the actor.
+        // Wait for the AnnotationLayer to get drawn before populating all the fields with data from the Document.
         pdfviewerapp.eventBus.on('annotationlayerrendered', (event) => {   // from PdfPageView
-            console.log(`annotationlayerrendered ${event.pageNumber}`);
-            if (!actor2pdfviewer.has(actorid)) actor2pdfviewer.set(actorid, event.source);
-            // Page has been rendered, so load the fields on the form with data from the Actor.
-            setFormFromActor(event.source, actor);
+            if (!document2pdfviewer.has(document.uuid)) document2pdfviewer.set(document.uuid, event.source);
+            setFormFromDocument(event.source, document);
         })
     })
 }
@@ -231,6 +244,11 @@ export async function initEditor(sheet, html, data) {
  * @param {Object} mapping 
  */
 export function registerActorMapping(mapping) {
-    pdf2actormap = mapping;
-    game.settings.set(PDFCONFIG.MODULE_NAME, PDFCONFIG.ACTOR_CONFIG, Obj2String(pdf2actormap));
+    map_pdf2actor = mapping;
+    game.settings.set(PDFCONFIG.MODULE_NAME, PDFCONFIG.ACTOR_CONFIG, Obj2String(map_pdf2actor));
+}
+
+export function registerItemMapping(mapping) {
+    map_pdf2item = mapping;
+    game.settings.set(PDFCONFIG.MODULE_NAME, PDFCONFIG.ACTOR_CONFIG, Obj2String(map_pdf2item));
 }